pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "brutils>=2.3.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
